+++ Package.swift
@@ -3,9 +3,6 @@ import PackageDescription
 
 let package = Package(
     name: "Clicore",
-    platforms: [
-        .macOS(.v10_15)
-    ],
     products: [
         .library(
             name: "Clicore",