+++ CHANGELOG.md
@@ -6,7 +6,9 @@
     - Correctly `nil` references to `DispatchSourceSignal` instances after cancellation.
     - Add a default error message on `runProcess(()` internal `Process.run()` immediate fails.
     - Abstract away `NSString` functions in `Path.swift` with a `String` extension.
+    - Make `Stdio.Settings` and `Stdio.Prefixes` initialisera public.
     - Add `Cli` module tests.
+    - Extend `Stdio` and `Path` test suites.
 - 0.5.0 *13 May 2026*
     - Add `Stdio.settings` struct for passing rendering settings.
     - Add emoji support to output as alternative to text.